+++ tests/test_clustering.py
@@ -1,6 +1,7 @@
 import numpy as np
 from numpy.testing import assert_array_equal
 import pytest
+from sklearn.base import clone
 from pytest import approx
 
 from pyriemann.clustering import (
@@ -279,6 +280,18 @@ def clt_score(clust, X, y=None):
     assert isinstance(score, float)
 
 
+@pytest.mark.parametrize("clust", clusts)
+def test_clustering_get_params(clust):
+    """Test sklearn compliance of get_params, set_params and clone"""
+    clt = clust()
+
+    params = clt.get_params()
+    assert clone(clt).get_params() == params
+
+    clt.set_params(metric="logeuclid")
+    assert clone(clt).get_params()["metric"] == "logeuclid"
+
+
 ###############################################################################
 
 
@@ -312,6 +325,14 @@ def test_kmeans(clust, init, n_init, metric, get_mats, get_labels):
         assert dists.shape == (n_matrices, n_clusters)
     elif clust is KmeansPerClassTransform:
         assert dists.shape == (n_matrices, clt.covmeans_.shape[0])
+
+
+def test_kmeansperclasstransform_get_params():
+    """Test that KmeansPerClassTransform exposes all Kmeans parameters"""
+    assert (
+        KmeansPerClassTransform().get_params().keys()
+        == Kmeans().get_params().keys()
+    )
 
 
 @np.vectorize